[Technical] Address review comments in Section 1 from David Noveck
> 1. Introduction
> In the first sentence, suggest "may employ" -> "provides for the employment of"
> In the second sentence, suggest, "must agree" --> "are bound by an agreement as to which"

No reason is given for these suggested substitutions. I'd like
to avoid wordy phrasing unless absolutely necessary, especially
in non-technical parts of this document.

Note that "may employ" comes directly from the Introduction of
RFC8267. Grammarly has no issue with that usage. But perhaps "can
employ" is more precise.

The second suggested replacement might be to avoid ambiguity with
the BCP14 keyword "MUST". However, the sentence is stating an
interoperability requirement. "MUST" is appropriate in this case.